Major reparameterization
+ Deprecate factorScalePrior as a user tweakable setting since it's
mostly useful to limit extreme proportions and the default should work
fine in most cases.

+ Increase corLKJPrior from 2.0 to 2.5 (divergences observed)

+ Increase propShape from 3.0 to 4.0 (divergences observed)

+ Change rawThreshold prior from beta(1.1, 1.1) to beta(1.1, 2).
This is probably just aesthetic because it worked fine before.

+ Change alpha prior from inv_gamma to normal(1.749, alphaScalePrior)
T[0,]. Alpha isn't a variance so inv_gamma is just wrong. With this new
prior, we can estimate alpha in the factor model (yay!).

+ In unidim_adapt, change the prior on sigma from the arbitrary
lognormal(1, 1) to a more suitable inv_gamma(1, 1). In addition, treat
it like a variance and use a normal(0, sqrt(sigma)) prior on theta. This
isn't much of an improvement in estimate accuracy, but seems to cure
some divergences and low_bfmi so it's a win!

+ Standardize resulting scores across all models.

+ Change rawUnique from beta(.) to uniform(-1,1) prior. Not sure if this
is optimal, but beta(.) just doesn't make a lot of sense.